add help 'srcs/to_help' add new function to fix cas where we try to print only '%'

+++ srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: aerrajiy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 20:38:08 by aerrajiy          #+#    #+#             */
-/*   Updated: 2022/10/24 15:56:12 by aerrajiy         ###   ########.fr       */
+/*   Updated: 2022/10/27 20:45:32 by aerrajiy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
